# Option selection speed improvements