--- a/Engine/DataFeeds/Enumerators/Factories/FuturesChainUniverseSubscriptionEnumeratorFactory.cs
+++ b/Engine/DataFeeds/Enumerators/Factories/FuturesChainUniverseSubscriptionEnumeratorFactory.cs
@@ -16,7 +16,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using QuantConnect.Data;
 using QuantConnect.Data.UniverseSelection;
 using QuantConnect.Interfaces;
@@ -32,15 +31,18 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
         private readonly bool _isLiveMode;
 
         private readonly IDataQueueUniverseProvider _symbolUniverse;
+        private readonly IDataCacheProvider _dataCacheProvider;
         private readonly ITimeProvider _timeProvider;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="FuturesChainUniverseSubscriptionEnumeratorFactory"/> class
         /// </summary>
         /// <param name="enumeratorConfigurator">Function used to configure the sub-enumerators before sync (fill-forward/filter/ect...)</param>
-        public FuturesChainUniverseSubscriptionEnumeratorFactory(Func<SubscriptionRequest, IEnumerator<BaseData>, IEnumerator<BaseData>> enumeratorConfigurator)
+        /// <param name="dataCacheProvider">The cache provider instance to use</param>
+        public FuturesChainUniverseSubscriptionEnumeratorFactory(Func<SubscriptionRequest, IEnumerator<BaseData>, IEnumerator<BaseData>> enumeratorConfigurator, IDataCacheProvider dataCacheProvider)
         {
             _isLiveMode = false;
+            _dataCacheProvider = dataCacheProvider;
             _enumeratorConfigurator = enumeratorConfigurator;
         }
 
@@ -67,36 +69,28 @@ namespace QuantConnect.Lean.Engine.DataFeeds.Enumerators.Factories
         {
             if (_isLiveMode)
             {
-                var subscriptionConfiguration = GetSubscriptionConfigurations(request).First();
-                var subscriptionRequest = new SubscriptionRequest(request, configuration: subscriptionConfiguration);
+                var subscriptionRequest = new SubscriptionRequest(request, configuration: GetSubscriptionConfiguration(request));
 
                 return new DataQueueFuturesChainUniverseDataCollectionEnumerator(subscriptionRequest, _symbolUniverse, _timeProvider);
             }
             else
             {
-                var factory = new BaseDataSubscriptionEnumeratorFactory(_isLiveMode);
+                var factory = new BaseDataSubscriptionEnumeratorFactory(_isLiveMode, _dataCacheProvider);
 
-                var enumerators = GetSubscriptionConfigurations(request)
-                    .Select(c => new SubscriptionRequest(request, configuration: c))
-                    .Select(sr => _enumeratorConfigurator(request, factory.CreateEnumerator(sr, dataProvider)));
+                var newRequest = new SubscriptionRequest(request, configuration: GetSubscriptionConfiguration(request));
+                var enumerator = _enumeratorConfigurator(request, factory.CreateEnumerator(newRequest, dataProvider));
 
-                var sync = new SynchronizingEnumerator(enumerators);
-                return new FuturesChainUniverseDataCollectionAggregatorEnumerator(sync, request.Security.Symbol);
+                return new FuturesChainUniverseDataCollectionAggregatorEnumerator(enumerator, request.Security.Symbol);
             }
         }
 
-        private static IEnumerable<SubscriptionDataConfig> GetSubscriptionConfigurations(SubscriptionRequest request)
+        private static SubscriptionDataConfig GetSubscriptionConfiguration(SubscriptionRequest request)
         {
             var config = request.Configuration;
             var resolution = config.Resolution;
 
-            var configurations = new List<SubscriptionDataConfig>
-            {
-                // rewrite the primary to be fill forward
-                new SubscriptionDataConfig(config, resolution: resolution, fillForward: true)
-            };
-
-            return configurations;
+            // rewrite the primary to be fill forward
+            return new SubscriptionDataConfig(config, resolution: resolution, fillForward: true);
         }
     }
 }