--- a/Engine/DataFeeds/SingleEntryDataCacheProvider.cs
+++ b/Engine/DataFeeds/SingleEntryDataCacheProvider.cs
@@ -14,11 +14,14 @@
  *
 */
 
+using System;
 using System.IO;
 using Ionic.Zip;
-using QuantConnect.Interfaces;
-using QuantConnect.Logging;
+using System.Linq;
 using QuantConnect.Util;
+using QuantConnect.Logging;
+using QuantConnect.Interfaces;
+using System.Collections.Generic;
 
 namespace QuantConnect.Lean.Engine.DataFeeds
 {
@@ -87,6 +90,22 @@ namespace QuantConnect.Lean.Engine.DataFeeds
         public void Store(string key, byte[] data)
         {
             //
+        }
+
+        /// <summary>
+        /// Returns a list of zip entries in a provided zip file
+        /// </summary>
+        public List<string> GetZipEntries(string zipFile)
+        {
+            var stream = _dataProvider.Fetch(zipFile);
+            if (stream == null)
+            {
+                throw new ArgumentException($"Failed to create source stream {zipFile}");
+            }
+            var entryNames = Compression.GetZipEntryFileNames(stream).ToList();
+            stream.DisposeSafely();
+
+            return entryNames;
         }
 
         /// <summary>