… Engine/DataFeeds/ZipEntryNameSubscriptionDataSourceReader.cs
… Tests/Engine/DataFeeds/Enumerators/Factories/BaseDataSubscriptionEnumeratorFactoryTests.cs
… Tests/Engine/DataFeeds/Enumerators/Factories/OptionChainUniverseSubscriptionEnumeratorFactoryTests.cs
… Tests/Engine/DataFeeds/FileSystemDataFeedTests.cs
… Tests/Engine/DataFeeds/IndexSubscriptionDataSourceReaderTests.cs
… Tests/Engine/DataFeeds/SubscriptionDataReaderTests.cs
… Tests/Engine/DataFeeds/TextSubscriptionDataSourceReaderTests.cs
… Tests/Engine/DataFeeds/ZipEntryNameSubscriptionFactoryTests.cs
… Tests/ToolBox/LeanDataReaderTests.cs