<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Allow caching zip entry names if possible for a performance improvement,
  Specially for daily and hourly futures and options.
- For option universe subscription request, non option data stack will not use `BaseDataSubscriptionEnumeratorFactory` but the normal data stack. Enables caching in some cases and minor bug fixes

#### PERFORMANCE

- Running on cloud B8-16

```csharp
    public class BasicTemplateOptionsDailyAlgorithm : QCAlgorithm
    {
        public override void Initialize()
        {
            SetStartDate(2020, 1, 1);
            SetEndDate(2020, 6, 1);

            UniverseSettings.Resolution = Resolution.Daily;
            Schedule.On(Schedule.DateRules.EveryDay(), Schedule.TimeRules.Noon, (_, _) =>
            {
                Plot("RAM", "RAM", (int)OS.ApplicationMemoryUsed);
            });
            AddUniverse(CoarseSelectionFunction);
        }

        public static IEnumerable<Symbol> CoarseSelectionFunction(IEnumerable<CoarseFundamental> coarse)
        {
            var sortedByDollarVolume = coarse.OrderByDescending(x => x.DollarVolume);
            var top = sortedByDollarVolume.Take(4);
            return top.Select(x => x.Symbol);
        }
        public override void OnSecuritiesChanged(SecurityChanges changes)
        {
            foreach (var security in changes.AddedSecurities)
            {
            	if(security.Symbol.SecurityType == SecurityType.Equity)
            	{
	                var option = AddOption(security.Symbol, Resolution.Daily);
	            	option.SetFilter(x =>
	            	{
	            		return x.CallsOnly().Strikes(0, 1).Expiration(0, 10);
	            	});
            	}
            }
        }
    }
```

#### MASTER

78.97 seconds at 285k data points per second. Processing total of 22,468,025 data points.
80.36 seconds at 280k data points per second. Processing total of 22,468,025 data points.
77.92 seconds at 288k data points per second. Processing total of 22,468,025 data points.

![image](https://user-images.githubusercontent.com/18473240/148557250-2b3b53d8-5096-49ae-b652-cd9c4a4fd080.png)

![image](https://user-images.githubusercontent.com/18473240/148556819-3f5842da-5f7b-453f-991e-6d0a6fdfc69e.png)

#### PR

29.45 seconds at 765k data points per second. Processing total of 22,518,355 data points.
28.84 seconds at 781k data points per second. Processing total of 22,518,355 data points.
31.11 seconds at 724k data points per second. Processing total of 22,518,355 data points.

![image](https://user-images.githubusercontent.com/18473240/148557971-4694caac-96d7-4fb6-a83a-6e7b581b3ca1.png)


![image](https://user-images.githubusercontent.com/18473240/148556856-676a252d-8e52-42d8-9f5b-9a63077b4f52.png)

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
N/A
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Speed improvements
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Cloud backtesting
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->